+++ rf_bk8500b/BK8500BLibrary/__init__.py
@@ -1,6 +1,6 @@
 """Robot Framework library for B&K Precision 8500B Series electronic loads."""
 from .library import BK8500BLibrary, BK8500BRobotError
 
-__version__ = "26.04"
+__version__ = "26.05"
 
 __all__ = ["BK8500BLibrary", "BK8500BRobotError"]